Generate RTSP session id with gcrypt as unsigned long, use value 0 as not valid.

+++ include/fenice/rtsp.h
@@ -66,7 +66,7 @@ typedef struct _RTSP_interleaved {
 
 typedef struct _RTSP_session {
     int cur_state;
-    int session_id;
+    unsigned long session_id;
     int started;
     RTP_session *rtp_session;
     struct _RTSP_session *next;
